utils: move test utilities to testutil
